Move card to Generation tab
